requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-currentuser==0.5.1
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.8.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
